Handle HTTP Unauthorized status
The http:get helper will now return an Err(ApiError::InvalidAccessToken)
when the http status code is set to StatusCode::Unauthorized. The
athlete stats api now returns an appropriate error instead of a json
decoding error when attempting to retrieve stats for a different athlete
than the currently authorized athlete.